Moving second part of PR#99 to section 3
Attempt at moving 

Further, it is out of scope for
this document if the old address can also be used for multipath
after a client has migrated to the address provided in the preferred_address
transport parameter. However, it SHOULD NOT be assumed that it is
possible to use both addresses simultaneously without further confirmation
from the other host.

In the beginning of section where I think it is more appropriate.