Remove mentions of the "Path Identifier"
In lot of places in the draft, we mention the "Path Identifier",
which we define as the Destination Connection ID sequence number
used over that path. It seems that maintaining the notion of
"Path Identifier" causes a lot of confusion, so better keep things
simple and mention CID sequence numbers directly.

Note that this implies that the related fields in
multipath-specific have been renamed to their "explicit" form. Some
of these names are maybe a bit long, so we may shorten them if this
does not introduce ambiguity.

Fix #169. Rewriting related section should also fix #181.