Editorial pass on explicit Path ID
I mainly merged the "Path identifier" section into the right place in the rest of the document. Having such a separate section did repeat the same things multiple time. That is usually very unfortunate because that means you have to change things in multiple places.

Further, some smaller editorial issue in the rest of the document.

This is only a first pass, I think we need more iterations.