Cleanup MAX_PATHS
There was a naming inconsistency here.

I went with one direction, which I think works best as it has a direct relation to the path identifier.  MAX_STREAMS uses a count because of the way in which streams are allocated odd/even.  That awkwardness we have for streams isn't necessary here.