Proposal to make path status sequence number per path.
fixed #312 

This is a proposal to make the sequence number in the path_avaiable and path_standby frames per-path. It is more inline with how the sequence number used in the mp_new_cid and mp_retire_cid frames. However, both approaches are fine. Please discuss in issue #312 which one you prefer.