TRANSPORT_PARAMETER_ERROR when initial_max_path is too large
fixes #432 

This PR defines an error for the case when the initial_max_path value exceeds 2^32-1. As we changes the transport parameter meaning with the introduction of the explicit path ID, it was an oversight to not define an error here. Only remaining question is if TRANSPORT_PARAMETER_ERROR is the right error type?