remove the term "active path" as it was not well defined
fixes #409

"active path" was used for both: a path that is open or actively used. This PR removed the word active and uses either "open" or "usable" to distinguish these cases.

If more appropriate it might also use the term "established" for an open path. Is that okay, or should we only use "open"?