Note that issuing CIDs for all paths might be beneficial
Based on the discussion in the interop call today, it was not clear why an endpoint would not want to issue CIDs for all path IDs up to the max path idea. If an endpoint wants to limit the number of paths, it should rather do it by limiting the max path ID explicitly. Therefore I propose this additional sentence to hind at doing this.