MUST handle migration
fixes #549 

This PR adds a MUST for handling migration as proposed by @martinthomson in the issue. I first tried to simple add his proposed sentence here but that didn't make sense as it would say the same thing twice. So maybe this works?

However, I personally am not convinced that this change is an improvement. We didn't want to repeat any MUSTs from RFC9000. So if we don't explicitly change it, those are still valid. Therefore we only added a MUST on the use of CID for the same path ID here. I really don't feel that new MUST here is necessary.