Make it clear the CIDs are not actually retired
The way this was worded suggested that the CIDs are really retired,
but that would involve sending PATH_RETIRE_CONNECTION_ID frames for
all CIDs of an abandoned path.  This is not required by § 3.4 which
only says you need to "treat the CIDs as immediately retired".

Aligning this wording here makes the intention clearer.